requirements.txt
Flask==2.3.3
Flask_Cors==4.0.0
flask_swagger_ui==4.11.1
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
pytest==7.4.2